pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.5.0",
    "google-cloud-aiplatform>=1.92.0",
    "google-cloud-storage>=2.19.0",
    "google-cloud-vision>=3.10.1",
    "openai>=1.78.0",
    "python-dotenv>=1.1.0",
]
